+++ control_state/LineTraceState.cpp
@@ -54,7 +54,7 @@ void LineTraceState::execute() {
 
 	/* 足の制御 */
 	// 前進値、旋回値を設定
-	turn = -this->pid->calc(this->lineMonitor->getBorderBrightness(),(double)this->lineMonitor->getBrightness(),-100,100);
+	turn = (int)-this->pid->calc(this->lineMonitor->getBorderBrightness(),this->lineMonitor->getBrightness(),-100,100);
 	// 足の制御実行
 	balancingWalker->setForwardTurn(forward, turn);
