+++ control_state/Test1State.cpp
@@ -58,7 +58,7 @@ void Test1State::execute() {
 
 	/* 足の制御 */
 	// 前進値、旋回値を設定
-	turn = -this->pid->calc(this->lineMonitor->getBorderBrightness(),(double)this->lineMonitor->getBrightness(),-100,100);	// サンプルコース
+	turn = (int)-this->pid->calc(this->lineMonitor->getBorderBrightness(),this->lineMonitor->getBrightness(),-100,100);	// サンプルコース
 	// 足の制御実行
 	balancingWalker->setForwardTurn(forward, turn);
